Fix clang 10.x warnings (#2518)
Note that some of the warnings related to the dpx/cineon readers highlight suspicious code, but I've opted to retain the existing behavior. The offending variables to not appear to actually do anything, they are just carried around as metadata.

+++ src/psd.imageio/psdinput.cpp
@@ -304,7 +304,7 @@ private:
     {
         // RGB = CompRGB - (1 - alpha) * Background;
         double scale = std::numeric_limits<T>::is_integer
-                           ? 1.0 / std::numeric_limits<T>::max()
+                           ? 1.0 / double(std::numeric_limits<T>::max())
                            : 1.0;
 
         for (; size; --size, data += nchannels)
@@ -323,7 +323,7 @@ private:
     {
         // RGB = (CompRGB - (1 - alpha) * Background) / alpha
         double scale = std::numeric_limits<T>::is_integer
-                           ? 1.0 / std::numeric_limits<T>::max()
+                           ? 1.0 / double(std::numeric_limits<T>::max())
                            : 1.0;
 
         for (; size; --size, data += nchannels)
@@ -345,7 +345,7 @@ private:
     void associateAlpha(T* data, int size, int nchannels, int alpha_channel)
     {
         double scale = std::numeric_limits<T>::is_integer
-                           ? 1.0 / std::numeric_limits<T>::max()
+                           ? 1.0 / double(std::numeric_limits<T>::max())
                            : 1.0;
         for (; size; --size, data += nchannels)
             for (int c = 0; c < nchannels; c++)
